Extract @Id mapping to a separate class
